chore(RBAC): Move all RBAC related methods into the RBAC module
